% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{calc_all_metrics}
\alias{calc_all_metrics}
\title{Todas las métricas a la vez: R², MAE, RMSE, MaxAE, IAE.}
\usage{
calc_all_metrics(x, S_emp, S_mod)
}
\arguments{
\item{x}{Rejilla de puntos \code{x} (creciente).}

\item{S_emp}{Supervivencia empírica sobre \code{x}.}

\item{S_mod}{Supervivencia modelada sobre \code{x}.}
}
\value{
Objeto de clase \code{PuigStats} con campos \code{R2}, \code{MAE},
\code{RMSE}, \code{MaxAE} e \code{IAE}.
}
\description{
Todas las métricas a la vez: R², MAE, RMSE, MaxAE, IAE.
}
