% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{calc_mae}
\alias{calc_mae}
\title{Error absoluto medio (MAE).}
\usage{
calc_mae(S_emp, S_mod)
}
\arguments{
\item{S_emp}{Supervivencia empírica.}

\item{S_mod}{Supervivencia modelada (misma longitud).}
}
\value{
Escalar \eqn{\ge 0}.
}
\description{
Error absoluto medio (MAE).
}
