% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{calc_maxae}
\alias{calc_maxae}
\title{Error absoluto máximo (MaxAE).}
\usage{
calc_maxae(S_emp, S_mod)
}
\arguments{
\item{S_emp}{Supervivencia empírica.}

\item{S_mod}{Supervivencia modelada (misma longitud).}
}
\value{
Escalar \eqn{\ge 0}.
}
\description{
Error absoluto máximo (MaxAE).
}
