% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Raíz del error cuadrático medio (RMSE).}
\usage{
calc_rmse(S_emp, S_mod)
}
\arguments{
\item{S_emp}{Supervivencia empírica.}

\item{S_mod}{Supervivencia modelada (misma longitud).}
}
\value{
Escalar \eqn{\ge 0}.
}
\description{
Raíz del error cuadrático medio (RMSE).
}
