% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highprec.R
\name{.kummer_1f1_decay}
\alias{.kummer_1f1_decay}
\title{Rama decreciente: e^{-y}·M(a, b, y) por expansión asintótica.}
\usage{
.kummer_1f1_decay(a, b, y, maxterms = 80)
}
\description{
Rama decreciente: e^{-y}·M(a, b, y) por expansión asintótica.
}
