% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highprec.R
\name{.kummer_M_stable_float}
\alias{.kummer_M_stable_float}
\title{Serie de Kummer convergente M(a, b, z) (doble precisión).}
\usage{
.kummer_M_stable_float(a, b, z, maxterms = 15000)
}
\description{
Serie de Kummer convergente M(a, b, z) (doble precisión).
}
