% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dpuig}
\alias{dpuig}
\title{Densidad de la distribución de Puig (convención \code{d/p/q/r}).}
\usage{
dpuig(x, lam, k, T, log = FALSE)
}
\arguments{
\item{x}{Puntos de evaluación (\eqn{x \ge 0}); puede ser vector.}

\item{lam}{Parámetro \eqn{\lambda \ge 0}.}

\item{k}{Dimensión efectiva continua (\eqn{k \ge 1}).}

\item{T}{Escala / precisión (\eqn{T > 0}).}

\item{log}{Si \code{TRUE}, devuelve \code{log(f(x))}.}
}
\value{
Densidad (o log-densidad).
}
\description{
Densidad de la distribución de Puig (convención \code{d/p/q/r}).
}
