% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_density.R
\name{MB_cumulative}
\alias{MB_cumulative}
\title{Acumulada del caso límite Maxwell–Boltzmann.}
\usage{
MB_cumulative(x, k, B)
}
\arguments{
\item{x}{Vector de puntos de evaluación (\eqn{x \ge 0}).}

\item{k}{Dimensión / grados de libertad reales (\eqn{k \ge 1}).}

\item{B}{Parámetro de escala (\eqn{B > 0}).}
}
\value{
Acumulada \eqn{F(x)} en \eqn{[0, 1]}.
}
\description{
\deqn{F(x) = 1 - S(x)}
}
