% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_density.R
\name{MB_surviving}
\alias{MB_surviving}
\title{Supervivencia del caso límite Maxwell–Boltzmann.}
\usage{
MB_surviving(x, k, B)
}
\arguments{
\item{x}{Vector de puntos de evaluación (\eqn{x \ge 0}).}

\item{k}{Dimensión / grados de libertad reales (\eqn{k \ge 1}).}

\item{B}{Parámetro de escala (\eqn{B > 0}).}
}
\value{
Supervivencia \eqn{S(x)} en \eqn{[0, 1]}.
}
\description{
\deqn{S(x) = \Gamma(k/2, x^2/(2B)) / \Gamma(k/2)}
(gamma incompleta superior regularizada).
}
