% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ppuig}
\alias{ppuig}
\title{Acumulada (o cola superior) de la distribución de Puig.}
\usage{
ppuig(q, lam, k, T, lower.tail = TRUE)
}
\arguments{
\item{q}{Puntos de evaluación (\eqn{q \ge 0}); puede ser vector.}

\item{lam}{Parámetro \eqn{\lambda \ge 0}.}

\item{k}{Dimensión efectiva continua (\eqn{k \ge 1}).}

\item{T}{Escala / precisión (\eqn{T > 0}).}

\item{lower.tail}{Si \code{TRUE} (por defecto) devuelve \eqn{F(q)};
si \code{FALSE} devuelve la supervivencia \eqn{S(q)}.}
}
\value{
Probabilidad en \eqn{[0, 1]}.
}
\description{
Acumulada (o cola superior) de la distribución de Puig.
}
