% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{Puig_ci996}
\alias{Puig_ci996}
\title{Intervalo central al 99.8\\%: \eqn{(P_{0.1}, P_{99.9})}.}
\usage{
Puig_ci996(lam, k = NULL, T = NULL)
}
\arguments{
\item{lam}{Parámetro \eqn{\lambda \ge 0}, o un objeto \code{PuigDistribution}
o \code{MB}.}

\item{k}{Dimensión efectiva continua (\eqn{k \ge 1}).}

\item{T}{Escala / precisión (\eqn{T > 0}).}
}
\value{
Vector con nombres \code{lo}/\code{hi}.
}
\description{
Intervalo central al 99.8\\%: \eqn{(P_{0.1}, P_{99.9})}.
}
