% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{Puig_stats}
\alias{Puig_stats}
\title{Estadísticos teóricos completos.}
\usage{
Puig_stats(lam, k = NULL, T = NULL)
}
\arguments{
\item{lam}{Parámetro \eqn{\lambda \ge 0}, o un objeto \code{PuigDistribution}
o \code{MB}.}

\item{k}{Dimensión efectiva continua (\eqn{k \ge 1}).}

\item{T}{Escala / precisión (\eqn{T > 0}).}
}
\value{
Lista con campos \code{mean}, \code{var}, \code{sig},
\code{skewness} y \code{kurtosis}.
}
\description{
Estadísticos teóricos completos.
}
