% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{Puig_std}
\alias{Puig_std}
\title{Desviación estándar teórica.}
\usage{
Puig_std(lam, k = NULL, T = NULL)
}
\arguments{
\item{lam}{Parámetro \eqn{\lambda \ge 0}, o un objeto \code{PuigDistribution}
o \code{MB}.}

\item{k}{Dimensión efectiva continua (\eqn{k \ge 1}).}

\item{T}{Escala / precisión (\eqn{T > 0}).}
}
\value{
\eqn{\sigma} (escalar).
}
\description{
\deqn{\sigma = \sqrt{\operatorname{Var}(X)}}
}
