% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qpuig}
\alias{qpuig}
\title{Cuantil de la distribución de Puig.}
\usage{
qpuig(p, lam, k, T)
}
\arguments{
\item{p}{Probabilidades en \eqn{[0, 1]} (escalar o vector).}

\item{lam}{Parámetro \eqn{\lambda \ge 0}.}

\item{k}{Dimensión efectiva continua (\eqn{k \ge 1}).}

\item{T}{Escala / precisión (\eqn{T > 0}).}
}
\value{
Cuantil \eqn{Q(p)}.
}
\description{
Cuantil de la distribución de Puig.
}
