% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{rpuig}
\alias{rpuig}
\title{Muestreo aleatorio de la distribución de Puig.}
\usage{
rpuig(n, lam, k, T)
}
\arguments{
\item{n}{Número de muestras.}

\item{lam}{Parámetro \eqn{\lambda \ge 0}.}

\item{k}{Dimensión efectiva continua (\eqn{k \ge 1}).}

\item{T}{Escala / precisión (\eqn{T > 0}).}
}
\value{
Vector de longitud \code{n} con \eqn{X \ge 0}.
}
\description{
Muestreo aleatorio de la distribución de Puig.
}
