% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_types.R
\name{[[.PuigFitResult}
\alias{[[.PuigFitResult}
\title{Desempaquetado tipo tupla: p, ind, st = Puig_fit(...)}
\usage{
\method{[[}{PuigFitResult}(x, i)
}
\description{
Desempaquetado tipo tupla: p, ind, st = Puig_fit(...)
}
